Move applyPatcher to its own file
