+++ staging/src/k8s.io/apiserver/pkg/endpoints/handlers/BUILD
@@ -38,6 +38,7 @@ go_test(
 go_library(
     name = "go_default_library",
     srcs = [
+        "apply.go",
         "create.go",
         "delete.go",
         "doc.go",
